Add configurable log fields; change default port.
- ADDITIONAL_LOG_FIELDS lets fields be added to every emitted log
message via config. The intended use case I have in mind is to add
both `container_app` to distinguish this app, as well as `parent_app`
as a reference to the task that is being monitored.

- Changing default port 9639 -> 9659, and actually claiming it on the
official prometheus wiki page. 9639 was claimed by a different project
in the mean time.